first split,not yet working
